[refactor] :recycle: rename keyGeneratorHelper a securityHelper (#276)
