+++ scripts/update_users_wallet_eoa.ts
@@ -8,7 +8,7 @@ import mongoose from 'mongoose';
 
 import { IUser } from '../src/models/user';
 import { Logger } from '../src/helpers/loggerHelper';
-import { generatePrivateKey } from '../src/helpers/keyGeneratorHelper';
+import { generatePrivateKey } from '../src/helpers/SecurityHelper';
 
 dotenv.config();
