+++ src/helpers/SecurityHelper.ts
